Allow server to change own nick
Previously, nick would only change once the client changes it with the
'/n' command (where it is stored temporarily in `_nick`) and the server
confirms the change with the NICK command (and _nick is copied to nick).

These changes make it so that nick only changes when the server says so,
regardless of what the client does. The '/n' command only sends a NICK
command to the server and nothing more. The server then replies (or
not).

This means that the server can change a user's nick without the user using the '/n' command. This prevents a lot of nasty side-effects that occur especially when using an IRC bouncer such as ZNC, where instantly upon connecting, the server will change your nick from e.g.:

bob/freenode             -> bob
bob@mobile/oftc          -> bobby
bob@mobile/oftc:password -> bob
etc.